Finalna prezetacija projekta

+++ Implementacija/pythonProject1/testpr1/filmkafe/views.py
@@ -21,8 +21,9 @@ from .serializers import MeniSerializer, ProductSerializer, CouponSerializer, Ev
     ProductSoldSerializer, WaiterWorkHoursSerializer, WaiterPermitSerializer
 from django.utils import timezone
 
+
 def hasPermit(user_id):
-    print("Date: "+str(datetime.datetime.today()))
+    print("Date: " + str(datetime.datetime.today()))
     return len(WaiterPermit.objects.filter(waiter=user_id).filter(day__gt=datetime.datetime.today())) > 0
 
 
@@ -123,7 +124,7 @@ def dogadjajiRezervacije(request):
         if request.user.role == 'Manager':
             temp = loader.get_template('filmkafe/DogadjajiRezervacije.html')
         elif request.user.role == 'Waiter':
-            if(hasPermit(request.user.idusers)):
+            if (hasPermit(request.user.idusers)):
                 temp = loader.get_template('filmkafe/DogadjajiRezervacije.html')
             else:
                 temp = loader.get_template('filmkafe/ZapocniSmenu.html')
@@ -213,6 +214,7 @@ def rasporedi(request):
     context = {'puf': puff}
     return HttpResponse(temp.render(context, request))
 
+
 def rasporedi_konobar(request):
     puff = ""
     if request.user.is_authenticated:
@@ -269,7 +271,7 @@ def skeniranjeQRKoda(request):
     puff = ""
     if request.user.is_authenticated:
         if request.user.role == 'Waiter':
-            if(hasPermit(request.user.idusers)):
+            if (hasPermit(request.user.idusers)):
                 temp = loader.get_template('filmkafe/SkeniranjeQR.html')
             else:
                 temp = loader.get_template('filmkafe/ZapocniSmenu.html')
@@ -330,6 +332,7 @@ def apiProduct(request):
         p = Product.objects.filter(amount__gt=F('minamount'))
         res = ProductSerializer(p, many=True)
         return Response(res.data)
+    return Response(status=403)
 
 
 @login_required
@@ -339,6 +342,7 @@ def apiProductAll(request):
         p = Product.objects.all()
         res = ProductSerializer(p, many=True)
         return Response(res.data)
+    return Response(status=403)
 
 
 @login_required
@@ -348,6 +352,7 @@ def apiDeficientProduct(request):
         p = Product.objects.filter(minamount__gte=F('amount'))
         res = ProductSerializer(p, many=True)
         return Response(res.data)
+    return Response(status=403)
 
 
 @login_required
@@ -373,6 +378,7 @@ def apiReservations(request):
         e = EventReservations.objects.all()
         res = EventReservationSerializer(e, many=True)
         return Response(res.data)
+    return Response(status=403)
 
 
 @login_required
@@ -432,8 +438,9 @@ def apiSetProduct(request):
             print("here")
             res.save()
             return Response("Success")
-            print(res.errors)
+        print(res.errors)
         return Response("Faliure")
+    return Response(status=403)
 
 
 @login_required
@@ -462,6 +469,29 @@ def apiUpdateProduct(request):
             print(prd)
 
         return Response("Success")
+    return Response(status=403)
+
+
+@login_required
+@api_view(['POST'])
+def apiProductPurchase(request):
+    if (request.user.role == 'Waiter' or request.user.role == 'Manager'):
+        p = request.data.copy()
+        p = p['purchase']
+        p = json.loads(p)
+        for each in p:
+            print(each)
+
+            amt = Product.objects.filter(name=each['name']).values('amount')
+            if len(amt) == 0:
+                continue
+            amt = amt[0]['amount']
+            amt = int(amt) + int(each['amount'])
+            prd = Product.objects.filter(name=each['name']).update(amount=amt)
+            print(prd)
+
+        return Response("Success")
+    return Response(status=403)
 
 
 @login_required
@@ -470,6 +500,7 @@ def apiDeleteProduct(request):
     if (request.user.role == 'Waiter' or request.user.role == 'Manager'):
         prd = Product.objects.filter(idproduct=request.data['idproduct']).delete()
         return Response("Success")
+    return Response(status=403)
 
 
 @api_view(['POST'])
@@ -518,6 +549,7 @@ def apiCreateWaiter(request):
             errcode = usr.save(usr)
         finally:
             return Response("Success")
+    return Response(status=403)
 
 
 @login_required
@@ -603,6 +635,7 @@ def apiCreateEvent(request):
                                       event=Events.objects.get(idevents=idevents))
         print(res.errors)
         return Response(res.data)
+    return Response(status=403)
 
 
 @login_required
@@ -619,6 +652,7 @@ def apiCreateCoupon(request):
         if (res.is_valid()):
             res.save()
         return Response(res.data)
+    return Response(status=403)
 
 
 @login_required
@@ -628,6 +662,7 @@ def apiDeleteEvent(request):
         id = request.data['idEvent']
         event = Events.objects.filter(idevents=id).delete()
         return Response()
+    return Response(status=403)
 
 
 @login_required
@@ -637,6 +672,7 @@ def apiDeleteCoupon(request):
         id = request.data['idcupon']
         event = Coupon.objects.filter(idcupon=id).delete()
         return Response()
+    return Response(status=403)
 
 
 @login_required
@@ -646,6 +682,7 @@ def apiWaiters(request):
         e = Users.objects.filter(role='Waiter')
         res = UsersSerializer(e, many=True)
         return Response(res.data)
+    return Response(status=403)
 
 
 @login_required
@@ -655,6 +692,7 @@ def apiDeleteWaiters(request):
         print(request.data)
         e = Users.objects.filter(idusers=request.data['idusers']).delete()
         return Response("success")
+    return Response(status=403)
 
 
 @login_required
@@ -677,8 +715,8 @@ def apiMeniAdd(request):
                 if (new.is_valid()):
                     new.save()
                 print(new.errors)
-
-    return Response("Success")
+        return Response("Success")
+    return Response(status=403)
 
 
 @login_required
@@ -689,6 +727,18 @@ def apiShift(request):
         smene = ShiftSerializer(sm, many=True)
         print(smene.data)
         return Response(smene.data)
+    return Response(status=403)
+
+
+@login_required
+@api_view(['GET'])
+def apiCWShift(request):
+    if (request.user.role == 'Waiter' or request.user.role == 'Manager'):
+        sm = Shift.objects.filter(waiter=request.user.idusers)
+        smene = ShiftSerializer(sm, many=True)
+        print(smene.data)
+        return Response(smene.data)
+    return Response(status=403)
 
 
 @login_required
@@ -718,6 +768,7 @@ def apiChangeShift(request):
 
         shifts = Shift.objects.filter(idshift__gt=i).delete()
         return Response('yes')
+    return Response(status=403)
 
 
 @login_required
@@ -725,7 +776,7 @@ def apiChangeShift(request):
 def apiSchedule(request):
     if (request.user.role == 'Waiter' or request.user.role == 'Manager'):
         # print(datetime.datetime.today())
-        date = datetime.datetime.today() - datetime.timedelta(days=1)
+        date = datetime.datetime.today() - datetime.timedelta(days=2)
         print("HELLO")
         print(str(date))
         print("HELLO")
@@ -733,6 +784,7 @@ def apiSchedule(request):
         smene = ScheduleSerializer(sm, many=True)
         print(smene.data)
         return Response(smene.data)
+    return Response(status=403)
 
 
 @login_required
@@ -740,10 +792,11 @@ def apiSchedule(request):
 def apiMySchedule(request):
     if (request.user.role == 'Waiter' or request.user.role == 'Manager'):
         print(datetime.datetime.today())
-        date = datetime.datetime.today() - datetime.timedelta(days=1)
+        date = datetime.datetime.today() - datetime.timedelta(days=2)
         sm = Schedule.objects.filter(day__gte=date).filter(waiter=request.user.idusers)
         smene = ScheduleSerializer(sm, many=True)
         return Response(smene.data)
+    return Response(status=403)
 
 
 @login_required
@@ -759,6 +812,7 @@ def apiChangeSchedule(request):
                 new.save()
             print(new.errors)
         return Response("sadness")
+    return Response(status=403)
 
 
 @login_required
@@ -779,6 +833,7 @@ def apiSetup(request):
         print(sets)
 
         return Response(sets)
+    return Response(status=403)
 
 
 @login_required
@@ -788,6 +843,7 @@ def apiTables(request):
         sp = Table.objects.all()
         smene = TableSerializer(sp, many=True)
         return Response(smene.data)
+    return Response(status=403)
 
 
 @login_required
@@ -817,6 +873,7 @@ def apiCreateTables(request):
         print("stop2\n")
         print(new.errors)
         return Response("failed")
+    return Response(status=403)
 
 
 @login_required
@@ -830,22 +887,33 @@ def apiCreateSetup(request):
             new.save()
             return Response("success")
         return Response("failed")
+    return Response(status=403)
 
 
 @login_required
 @api_view(['POST'])
 def apiDeleteSetup(request):
     if (request.user.role == 'Manager'):
-        e = Setup.objects.filter(idsetup=request.data['idsetup']).delete()
-        return Response("success")
+        try:
+            e = Setup.objects.filter(idsetup=request.data['idsetup']).delete()
+            return Response("success")
+        except:
+            return Response("failure")
+    return Response(status=403)
 
 
 @login_required
 @api_view(['POST'])
 def apiDeleteTables(request):
     if (request.user.role == 'Manager'):
+        setupTable = SetupTables.objects.filter(table=Table.objects.get(pk=request.data['idtable']))
+        print(setupTable)
+        events = Events.objects.filter(setup=setupTable[0].setup)
+        if len(events)>0:
+            return Response("failure")
         e = Table.objects.filter(idtable=request.data['idtable']).delete()
         return Response("Success")
+    return Response(status=403)
 
 
 @login_required
@@ -855,6 +923,7 @@ def apiPreference(request):
         e = Preferences.objects.all()
         res = PreferenceSerializer(e, many=True)
         return Response(res.data)
+    return Response(status=403)
 
 
 @api_view(['GET'])
@@ -863,6 +932,7 @@ def apiMyPreference(request):
         e = Preferences.objects.filter(waiter=request.user.idusers)
         res = PreferenceSerializer(e, many=True)
         return Response(res.data)
+    return Response(status=403)
 
 
 @login_required
@@ -885,6 +955,7 @@ def apiSetPreference(request):
                     ser.save()
                 print(ser.errors)
         return Response("")
+    return Response(status=403)
 
 
 @login_required
@@ -896,6 +967,7 @@ def apiCustomerExpeneture(request):
         curr = Users.objects.filter(idusers=dat['idusers']).values('salary')
         user = Users.objects.filter(idusers=dat['idusers'])
         exp = Users.objects.filter(idusers=dat['idusers']).values('totalexp')
+        print(curr)
         curr = curr[0]['salary'] + int(dat['expendature'])
         exp = exp[0]['totalexp'] + int(dat['expendature'])
         print(curr)
@@ -914,6 +986,7 @@ def apiCustomerExpeneture(request):
             e.save()
         print(e.errors)
         return Response('res.data')
+    return Response(status=403)
 
 
 @login_required
@@ -926,7 +999,7 @@ def apiGetCustomerExpeneture(request):
         res = CustomerExpenditureSerializer(expendatures, many=True)
         return Response(res.data)
     else:
-        return Response("FORBIDEN")
+        return Response(status=403)
 
 
 @login_required
@@ -939,7 +1012,7 @@ def apiGetProductSold(request):
         res = ProductSoldSerializer(prodSold, many=True)
         return Response(res.data)
     else:
-        return Response("FORBIDEN")
+        return Response(status=403)
 
 
 @login_required
@@ -961,6 +1034,7 @@ def apiReservationsNA(request):
             each['eventid'] = each['event']
             each['idusers'] = each['user']
         return Response(reservations)
+    return Response(status=403)
 
 
 @login_required
@@ -981,12 +1055,13 @@ def apiReservationsA(request):
             rt = ReservedTablesNew.objects.filter(reservation=each['id'])
             each['tableIds'] = ReservedTablesSerializer(rt, many=True).data
         return Response(reservations)
+    return Response(status=403)
 
 
 @login_required
 @api_view(['GET'])
 def apiFreeTables(request):
-    if (request.user.role == 'Waiter' or request.user.role == 'Manager'):
+    if request.user.role == 'Waiter' or request.user.role == 'Manager':
         events = Events.objects.all()
         ev = EventsSerializer(events, many=True)
         ev = ev.data
@@ -997,6 +1072,7 @@ def apiFreeTables(request):
                 tables.append(tab['table'])
             each['freeTableIds'] = tables
         return Response(ev)
+    return Response(status=403)
 
 
 @api_view(['GET'])
@@ -1011,7 +1087,7 @@ def apiCheckLogInUser(request):
 @login_required
 @api_view(['POST'])
 def apiStart(request):
-    if (request.user.role == 'Waiter'):
+    if (request.user.role == 'Waiter' or request.user.role == 'Manager'):
         print(request.data)
         if (request.user.is_authenticated):
             time_change = datetime.timedelta(hours=24)
@@ -1024,11 +1100,13 @@ def apiStart(request):
                                           day__lt=after).update(started=datetime.datetime.now())
             print(before)
         return Response("Success")
+    return Response(status=403)
+
 
 @login_required
 @api_view(['POST'])
 def apiStartSchedule(request):
-    if (request.user.role == 'Waiter'):
+    if (request.user.role == 'Waiter' or request.user.role == 'Manager'):
         data = request.data.copy()
         print(data)
         schedule = Schedule.objects.filter(idschedule=data["idschedule"])[0]
@@ -1036,12 +1114,13 @@ def apiStartSchedule(request):
         schedule.ended = data["ended"]
         schedule.save()
         return Response("Success")
-    return Response("Failure")
+    return Response(status=403)
+
 
 @login_required
 @api_view(['POST'])
 def apiEnd(request):
-    if (request.user.role == 'Waiter'):
+    if (request.user.role == 'Waiter' or request.user.role == 'Manager'):
         if (request.user.is_authenticated):
             time_change = datetime.timedelta(hours=24)
             before = datetime.datetime.today() - time_change
@@ -1050,21 +1129,19 @@ def apiEnd(request):
                                           day__lt=after).update(ended=datetime.datetime.now())
             print(request.user.idusers)
         return Response("Success")
-
-
-
+    return Response(status=403)
 
 
 @login_required
 @api_view(['POST'])
 def apiReserve(request):
-    if (request.user.is_authenticated):
+    if request.user.is_authenticated:
         dat = request.data.copy()
         dat['approved'] = 0
         dat['user'] = request.user.idusers
         dat['event'] = dat['id']
         new = EventReservationSerializer(data=dat)
-        if (new.is_valid()):
+        if new.is_valid():
             new.save()
         print(new.errors)
     return Response("Success")
@@ -1073,7 +1150,7 @@ def apiReserve(request):
 @login_required
 @api_view(['POST'])
 def apiPostWaiterWorkHours(request):
-    if request.user.role == 'Waiter':
+    if (request.user.role == 'Waiter' or request.user.role == 'Manager'):
         data = request.data.copy()
         waiterHours = {}
         waiterHours["day"] = data['day']
@@ -1085,7 +1162,7 @@ def apiPostWaiterWorkHours(request):
         if res.is_valid():
             res.save()
         return Response("")
-    return Response("FORBIDDEN")
+    return Response(status=403)
 
 
 @login_required
@@ -1098,13 +1175,13 @@ def apiGetWaiterWorkHours(request):
         res = WaiterWorkHoursSerializer(objects, many=True)
         return Response(res.data)
     else:
-        return Response("FORBIDEN")
+        return Response(status=403)
 
 
 @login_required
 @api_view(['POST'])
 def apiPostWaiterPermit(request):
-    if request.user.role == 'Waiter':
+    if (request.user.role == 'Waiter' or request.user.role == 'Manager'):
         data = request.data.copy()
         waiterPermit = {}
         waiterPermit["day"] = data['day']
@@ -1114,20 +1191,18 @@ def apiPostWaiterPermit(request):
         if res.is_valid():
             res.save()
         return Response("")
-    return Response("FORBIDDEN")
+    return Response(status=403)
 
 
 @login_required
 @api_view(['GET'])
 def apiGetWaiterPermit(request):
-    if request.user.role == 'Waiter':
+    if (request.user.role == 'Waiter' or request.user.role == 'Manager'):
         permit = hasPermit(request.user.idusers)
         if permit:
             return Response("True")
         return Response("False")
-    return Response("FORBIDDEN")
-
-
+    return Response(status=403)
 
 
 @login_required
@@ -1147,29 +1222,33 @@ def apiApproveReservation(request):
         tables = json.loads(request.data['tables'])
         for each in tables:
             tab = {}
+            print(each)
             FreeTables.objects.filter(table=each['idtable'], event=request.data['idevents']).delete()
             tab['reservation'] = res[0].id
             tab['reservedtables'] = each['idtable']
             rt = ReservedTablesSerializer(data=tab)
+            print(rt.is_valid())
             if (rt.is_valid()):
                 rt.save()
             print(rt.errors)
 
         return Response("Success")
+    return Response(status=403)
 
 
 @login_required
 @api_view(['POST'])
 def apiDenyReservation(request):
-    if (request.user.role == 'Waiter'):
+    if (request.user.role == 'Waiter' or request.user.role == 'Manager'):
         res = EventReservations.objects.filter(user=request.data['idusers'], event=request.data['idevents']).delete()
         return Response("Success")
+    return Response(status=403)
 
 
 @login_required
 @api_view(['POST'])
 def apiBuyCupon(request):
-    if (request.user.role == 'Waiter'):
+    if (request.user.role == 'Waiter' or request.user.role == 'Manager'):
         print(request.data)
         cpn = Coupon.objects.filter(idcupon=request.data['idcoupon'])
         print(cpn)
@@ -1181,6 +1260,7 @@ def apiBuyCupon(request):
             usr.update(salary=(usrpts - pts))
             return Response(CouponSerializer(cpn).data)
         return Response({'idcupon': -1, 'description': "Nema dovoljno poena", 'name': 'Neuspeh'})
+    return Response(status=403)
 
 
 '''d=res.data
